Convert MobileUnit to interface
A Worker must be mobile and the relevent methods are duplicated in the MobileUnit class

Changes:
- Convert MobileUnit to AbstractMobileUnit
- Change the direct subclasses to use the new Abstract class
- Create a new MobileUnit interface that holds the relevent movement methods
- Worker interface extends the MobileUnit interface

Part of #1466
