+++ mars-sim-core/src/main/java/com/mars_sim/core/person/Person.java
@@ -79,7 +79,7 @@ import com.mars_sim.core.time.ClockPulse;
 import com.mars_sim.core.time.MarsTime;
 import com.mars_sim.core.time.Temporal;
 import com.mars_sim.core.tool.RandomUtil;
-import com.mars_sim.core.unit.MobileUnit;
+import com.mars_sim.core.unit.AbstractMobileUnit;
 import com.mars_sim.core.vehicle.Crewable;
 import com.mars_sim.core.vehicle.Rover;
 import com.mars_sim.core.vehicle.Vehicle;
@@ -88,7 +88,7 @@ import com.mars_sim.core.vehicle.Vehicle;
  * The Person class represents a person on Mars. It keeps track of everything
  * related to that person and provides information about him/her.
  */
-public class Person extends MobileUnit implements Worker, Temporal, Appraiser {
+public class Person extends AbstractMobileUnit implements Worker, Temporal, Appraiser {
 
 	/** default serial id. */
 	private static final long serialVersionUID = 1L;
