+++ mars-sim-core/src/main/java/com/mars_sim/core/unit/MobileUnit.java
@@ -1,15 +1,5 @@
-/*
- * Mars Simulation Project
- * MobileUnit.java
- * @date 2024-09-15
- * @author Barry Evans
- */
 package com.mars_sim.core.unit;
 
-import com.mars_sim.core.Unit;
-import com.mars_sim.core.UnitEventType;
-import com.mars_sim.core.UnitType;
-import com.mars_sim.core.location.LocationStateType;
 import com.mars_sim.core.map.location.Coordinates;
 import com.mars_sim.core.map.location.LocalPosition;
 import com.mars_sim.core.structure.Settlement;
@@ -17,242 +7,76 @@ import com.mars_sim.core.structure.building.Building;
 import com.mars_sim.core.vehicle.Vehicle;
 
 /**
- * Represent a Unit that can move around the surface
+ * Represents an entity that can be mobile.
  */
-public abstract class MobileUnit extends Unit  {
+public interface MobileUnit {
+	/**
+	 * Gets the coordinates on Mars surface.
+	 * 
+	 * @return
+	 */
+	public Coordinates getCoordinates();
+
+	/**
+	 * Is the worker in a vehicle ?
+	 *
+	 * @return true if the worker in a vehicle
+	 */
+	public boolean isInVehicle();
+
+	/**
+	 * Gets vehicle worker is in, null if member is not in vehicle/
+	 *
+	 * @return the worker's vehicle
+	 */
+	public Vehicle getVehicle();
+    	
+	/**
+	 * Is the worker in a settlement ?
+	 *
+	 * @return true if the worker in a settlement
+	 */
+	public boolean isInSettlement();
     
-    private Settlement owner;
-    private LocalPosition localPosn = LocalPosition.DEFAULT_POSITION;
-    private int currentBuildingInt;
-    private Coordinates surfacePosn;
+	/**
+	 * Gets the current Settlement of the worker; may be different from the associated Settlement.
+	 * 
+	 * @return
+	 */
+	public Settlement getSettlement();
 
     /**
-	 * Constructor.
+	 * Gets the Worker's building.
 	 * 
-	 * @param name the name of the unit
-	 * @param owner the unit's location
+	 * @return building
 	 */
-	protected MobileUnit(String name, Settlement owner) {
-		super(name, owner.getCoordinates()); // TODO Unit constructor needs to drop coordinates from params
-        this.surfacePosn = owner.getCoordinates();
-        this.owner = owner;
-
-		setContainer(owner);
-	}
-
+	public Building getBuildingLocation();
+	
 	/**
-	 * Set the container of this mobile unit
-	 * @param destination New destination of container
-	 */
-	protected void setContainer(Unit destination) {
-		setContainerID(destination.getIdentifier());
-	}
-
-	/**
-	 * Gets the position at a settlement.
-	 *
-	 * @return distance (meters) from the settlement's center.
-	 */
-	public LocalPosition getPosition() {
-		return localPosn;
-	}
-
-	/**
-	 * Sets the settlement-wide position.
+	 * Sets the building the worker is located at.
 	 *
 	 * @param position
 	 */
-	public void setPosition(LocalPosition position) {
-		this.localPosn = position;
-	}
-
-    /**
-	 * Computes the building the person is currently located at.
-	 * Returns null if outside of a settlement.
-	 *
-	 * @return building
-	 */
-	@Override
-	public Building getBuildingLocation() {
-		if (currentBuildingInt == -1)
-			return null;
-		return unitManager.getBuildingByID(currentBuildingInt);
-	}
+	public void setCurrentBuilding(Building building);
 
 	/**
-	 * Computes the building the person is currently located at.
-	 * Returns null if outside of a settlement.
-	 *
-	 * @return building
+	 * Gets the Worker's position within the Settlement/Vehicle.
+	 * 
+	 * @return
 	 */
-	public void setCurrentBuilding(Building building) {
-		if (building == null) {
-			currentBuildingInt = -1;
-		}
-		else {
-			currentBuildingInt = building.getIdentifier();
-		}
-	}
-
-
-	/**
-	 * Gets the unit's location.
-	 *
-	 * @return the unit's location
-	 */
-	public Coordinates getCoordinates() {
-		Unit cu = getContainerUnit();
-		if (cu.getUnitType() == UnitType.MARS) {	
-			// Since Mars surface has no coordinates, 
-			// Get from its previously setting location
-			return surfacePosn;
-		}
-		
-		// Unless it's on Mars surface, get its container unit's coordinates
-		return cu.getCoordinates();
-	}
-
-	/**
-	 * Sets unit's location coordinates.
-	 *
-	 * @param newLocation the new location of the unit
-	 */
-	public void setCoordinates(Coordinates newLocation) {
-        // TODO the null check can be removed once Unit stop managing coordinates
-		if ((surfacePosn == null) || !surfacePosn.equals(newLocation)) {
-			surfacePosn = newLocation;
-			fireUnitUpdate(UnitEventType.LOCATION_EVENT, newLocation);
-		}
-	}
-    
-    /**
-	 * This method assumes the Unit could be movable and change container. It identifies the
-	 * appropriate container it can use. Ideally this method should be moved to a 
-	 * new subclass called 'MovableUnit' that encapsulates some positioning methods 
-	 * not applicable to Structures.
-	 */
-	public String getContext() {
-		// Check vehicle first because could be in a Vehicle in a Settlement
-		if (isInVehicle()) {
-			return getVehicle().getChildContext();
-		}
-		else if (isInSettlement()) {
-			var b = getBuildingLocation();
-			if (b != null) {
-				return b.getContext();
-			}
-			else {
-				return getAssociatedSettlement().getName();
-			}
-		}
-		else if (isOutside()) {
-			return getCoordinates().getFormattedString();
-		}
-		else {
-			return getContainerUnit().getName();
-		}
-	}
-
-    /**
-     * Settlement that is associated with this Unit.
-     */
-    @Override
-    public Settlement getAssociatedSettlement() {
-        return owner;
-    }   
-    
-    /**
-	 * Gets the settlement the Unit is at.
-	 * Returns null if unit is not at a settlement.
-	 *
-	 * @return the settlement
-	 */
-	public Settlement getSettlement() {
-
-		if (getContainerID() <= Unit.MARS_SURFACE_UNIT_ID)
-			return null;
-
-		var c = getContainerUnit();
-
-		if (c instanceof Settlement s) {
-			return s;
-		}
-
-		else if (c instanceof Vehicle v) {
-			// Will see if vehicle is inside a garage or not
-			return (v.isInVehicleInGarage() ? v.getSettlement() : null);
-		}
-
-		else if (c instanceof FixedUnit b) {
-			return b.getAssociatedSettlement();
-		}
-		else if (c instanceof MobileUnit m) {
-			return m.getSettlement();
-		}
-
-		return null;
-	}
-
-    /**
-	 * Is this unit inside a settlement ?
-	 *
-	 * @return true if the unit is inside a settlement
-	 */
-	@Override
-	public boolean isInSettlement() {
-		return (getSettlement() != null);
-	}
-
-    
-	/**
-	 * Gets vehicle person is in, null if person is not in vehicle.
-	 *
-	 * @return the person's vehicle
-	 */
-	@Override
-	public Vehicle getVehicle() {
-		if (getLocationStateType() == LocationStateType.INSIDE_VEHICLE) {
-			return (Vehicle) getContainerUnit();
-		}
-
-		return null;
-	}
-    
-	/**
-	 * Is this unit inside a vehicle ?
-	 *
-	 * @return true if the unit is inside a vehicle
-	 */
-	public boolean isInVehicle() {
-		if (LocationStateType.INSIDE_VEHICLE == currentStateType)
-			return true;
-
-		if (LocationStateType.ON_PERSON_OR_ROBOT == currentStateType)
-			return getContainerUnit().isInVehicle();
-
-		return false;
-	}
-
+	public LocalPosition getPosition();
 	
-    /**
-	 * Updates the location state type of a person.
+	/**
+	 * Sets the worker's position at a settlement.
 	 *
-	 * @param newContainer
+	 * @param position
 	 */
-	protected void updateLocationState(Unit newContainer) {
-		if (newContainer == null) {
-			currentStateType = LocationStateType.UNKNOWN;
-			return;
-		}
+	public void setPosition(LocalPosition position);
 
-		currentStateType = switch (newContainer.getUnitType()) {
-            case SETTLEMENT -> LocationStateType.INSIDE_SETTLEMENT;
-            case BUILDING -> LocationStateType.INSIDE_SETTLEMENT;
-            case VEHICLE -> LocationStateType.INSIDE_VEHICLE;
-            case CONSTRUCTION -> LocationStateType.MARS_SURFACE;
-            case PERSON -> LocationStateType.ON_PERSON_OR_ROBOT;
-            case MARS -> LocationStateType.MARS_SURFACE;
-            default -> null;
-        };
-	}
+	/**
+	 * Is the Worker outside ?
+	 *
+	 * @return true if the worker is on the MarsSurface
+	 */
+	public boolean isOutside();
 }
